upgrade clickhouse to v26.7.2.59-stable
Raw 3-way sync of upstream ClickHouse v26.7.2.59-stable onto the
v26.5.1.882-stable baseline, same scope filter as the previous upgrades:
chdb-retained paths only, deliberately-dropped paths excluded.

For chdb-modified files that upstream also changed (167 files), the
upstream tag content is staged here; chdb-specific patches are re-applied
in the follow-up fixups commit to keep this commit bisectable.

Submodules: 35 tracked pins bumped to v26.7 values; 8 new submodules
added (libdeflate, mapbox-geometry, musl, simdutf, sz3, wagyu, xsimd,
zxc); base64 and curl removed (dropped upstream); chdb-custom pins left
untouched.

+++ src/Processors/QueryPlan/BuildQueryPipelineSettings.cpp
@@ -31,12 +31,13 @@ BuildQueryPipelineSettings::BuildQueryPipelineSettings(ContextPtr from)
     aggregation_memory_efficient_merge_threads = from->getSettingsRef()[Setting::aggregation_memory_efficient_merge_threads];
     min_outstreams_per_resize_after_split = from->getSettingsRef()[Setting::min_outstreams_per_resize_after_split];
     max_streams_for_union_step = from->getSettingsRef()[Setting::max_streams_for_union_step];
-    max_streams_for_union_step_to_max_threads_ratio = from->getSettingsRef()[Setting::max_streams_for_union_step_to_max_threads_ratio];
+    max_streams_for_union_step_to_max_threads_ratio = static_cast<double>(from->getSettingsRef()[Setting::max_streams_for_union_step_to_max_threads_ratio]);
 
     /// Setting query_plan_merge_filters is enabled by default.
     /// But it can brake short-circuit without splitting filter step into smaller steps.
     /// So, enable and disable this optimizations together.
     enable_multiple_filters_transforms_for_and_chain = settings[Setting::query_plan_merge_filters];
+
     block_marshalling_callback = from->getBlockMarshallingCallback();
 }
